+++ src/config/constatns.jsx
@@ -1,0 +1,30 @@
+import { Dimensions } from "react-native";
+
+export const WINDOWS = {
+  HEIGHT: Dimensions.get("window").height,
+};
+
+export const COORDINATES = {
+  END_POSITION_Y: -340,
+  FIRST_WEEK_Y: 207,
+  CALENDAR_MATRIX_HEIGHT_Y: 270,
+  DIVISION_LINE_Y: 450,
+};
+
+export const COLORS = {
+  TRANSPARENCY: "#00000000",
+  BLACK: "#000000",
+  WHITE: "#FFFFFF",
+  LIGHT_GREY: "#EEEEEE",
+  GREY: "#9E9E9E",
+  SKY_BLUE: "#35C0F5",
+};
+
+export const FONT_SIZE = {
+  MIDDLE: 15,
+};
+
+export const DAY_SIZE = {
+  WIDTH: 45,
+  HEIGHT: 45,
+};
